Re-engineered with react
This was a masters dissertertation project where I re-engineered MonkeyPuzzle using the React library.  This was to invesigate the benefits of maintainability to the software.  The final version was then saved as a design document to an instance of CouchDB where it could be served to the user as a CouchApp.